+++ NEWS.md
@@ -2,6 +2,7 @@
 
 * Adapting `sd_sigma` argument in `fitSlope` based on scale of the
   posterior coefficients for B, `beta_hat_b`.
+* Also output priors from `fitSlope`.
 
 # mrlocus 0.0.14